pyproject.toml
[tool.poetry]
name = "pytmm"
version = "0.1.0"
description = "PYthon TileMap Manager is a small command line program to manage your tilemaps."
authors = ["ambertide <egeemirozkan24@gmail.com>"]
readme = "README.md"
license = "MPL-2.0"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.0.1"
click = "^8.1.7"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
